move tests of `join` into `join/test.rs`
